Rewrite everything
lol just kidding, I just copy-pasted a bit of code around and made the
server code more robust. The service now just runs two threads which
communicate via an in-memory queue and both try to reconnect infinitely
if either mpd, mqtt or both goes down.

There was no good way to use mpd idle with a timeout (a timeout is
always a good idea), so this implementation now uses blocking polling
with a timeout instead. The service will terminate cleanly when SIGINT
is sent.